ducktypemacro rough draft
