+++ src/DuckDispatch.jl
@@ -16,7 +16,6 @@ end
     macro test_throws end
 end
 
-using BangBang: append!!#, push!!
 # using Base: tail
 using TestItems: @testitem
 using SumTypes: @sum_type, @cases
@@ -28,6 +27,7 @@ include("Types.jl")
 include("TypeUtils.jl")
 include("BehaviorDispatch.jl")
 include("MethodDispatch.jl")
+include("DuckTypeMacro.jl")
 
 @testitem "Test Basics" begin
     # Content Generated by @duck_type macro
@@ -84,16 +84,16 @@ include("MethodDispatch.jl")
         end
 
         # Here is an example of DuckType created by composing new behaviors with Iterable
-        struct IsContainer{T} <: DuckDispatch.DuckType{
-            Union{
-                DuckDispatch.Behavior{typeof(length), Tuple{DuckDispatch.This}},
-                DuckDispatch.Behavior{typeof(getindex), Tuple{DuckDispatch.This, Int}}
-            },
-            Union{
-                Iterable{T}
-            }
-        }
-        end
+        # struct IsContainer{T} <: DuckDispatch.DuckType{
+        #     Union{
+        #         DuckDispatch.Behavior{typeof(length), Tuple{DuckDispatch.This}},
+        #         DuckDispatch.Behavior{typeof(getindex), Tuple{DuckDispatch.This, Int}}
+        #     },
+        #     Union{
+        #         Iterable{T}
+        #     }
+        # }
+        # end
         # The narrow function is used to find the most specific DuckType that can wrap a given object
         function DuckDispatch.narrow(::Type{<:IsContainer}, ::Type{T}) where {T}
             E = eltype(T)
@@ -133,6 +133,11 @@ include("MethodDispatch.jl")
     @test iterate(DuckDispatch.wrap(Iterable{Int}, [1, 2, 3])) == (1, 2)
     @test iterate(DuckDispatch.wrap(IsContainer{Int}, [1, 2, 3])) == (1, 2)
     @test length(DuckDispatch.wrap(IsContainer{Int}, [1, 2, 3])) == 3
+
+    DuckDispatch.@duck_type struct IsContainer2{T} <: Union{IsIterable{T}}
+        function length(::DuckDispatch.This)::Int end
+        function getindex(::DuckDispatch.This, ::Int)::T end
+    end
 
     DuckDispatch.@duck_dispatch function collect_ints(arg1::IsContainer{T}) where {T}
         return T[x for x in arg1]
